Add spillable map and resolve map close
